+++ src/index.ts
@@ -69,6 +69,13 @@ const defaultLogger = (options: AckeeLoggerOptions & { loggerName?: string } = {
     serializers.disablePaths(options.disableFields);
     serializers.enablePaths(options.enableFields);
 
+    if (options.sentry) {
+        const sentry = require('@sentry/node');
+        if (typeof options.sentry === 'string') {
+            sentry.init({ dsn: options.sentry });
+        }
+    }
+
     const isTesting = process.env.NODE_ENV === 'test';
     const defaultLevel: Level = options.defaultLevel || (isTesting ? 'silent' : 'debug');
     const messageKey = 'message'; // best option for Google Stackdriver,