+++ src/streams.ts
@@ -6,6 +6,7 @@ import * as util from 'util';
 import { loggerNameKey, pkgVersionKey } from '.';
 import { AckeeLoggerOptions, AckeeLoggerStream } from './interfaces';
 import { levels } from './levels';
+import { SentryTransformStream } from './sentry';
 import { StackDriverFormatStream } from './stackdriver';
 
 const pkgJson = JSON.parse(fs.readFileSync(path.resolve(path.join(__dirname, '..', 'package.json')), 'utf8'));
@@ -71,6 +72,11 @@ const initLoggerStreams = (
     }
 
     streams = decorateStreams(streams, getDefaultTransformStream(options));
+
+    if (options.sentry) {
+        streams = decorateStreams(streams, SentryTransformStream);
+    }
+
     return streams;
 };